Break cyclic dependencies in `libparsec_protocol`
The crate `libparsec_protocol` depends on `libparsec_serialization_format` which it self depends on `libparsec_protocol::Request` (in dev-dependencies for testing purpose).

To resolve that, I've create an intermediate crate `libparsec_miniprotocol` that will be used by the crate `protocol` and `serialization_format`.

Others Changes
--------------

- Updated tests to use `libparsec_miniprotocol` when needed.
- Rework how we expose `libparsec::protocol`.
- Update `serialization_format` to use `libparsec_miniprotocol` when generating code.
