+++ oxidation/libparsec/Cargo.toml
@@ -32,6 +32,7 @@ libparsec_testbed = { path = "crates/testbed", optional = true }
 # TODO: merge with regular dependencies
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 libparsec_protocol = { path = "crates/protocol" }
+libparsec_miniprotocol = { path = "crates/miniprotocol" }
 libparsec_core_fs = { path = "crates/core_fs" }
 libparsec_types = { path = "crates/types" }
 # TODO: should not be an optional
