+++ oxidation/libparsec/crates/miniprotocol/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "libparsec_miniprotocol"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rmp-serde = "1.1.1"
+serde = { version = "1.0.147", features = ["derive"], default-features = false }
