+++ oxidation/libparsec/crates/protocol/src/lib.rs
@@ -3,12 +3,7 @@
 mod error;
 mod handshake;
 
-use serde::{Deserialize, Serialize};
-use std::num::NonZeroU8;
-
 use libparsec_serialization_format::parsec_protocol_cmds_family;
-
-use crate as libparsec_protocol;
 
 pub use error::*;
 pub use handshake::*;
@@ -28,26 +23,6 @@ macro_rules! impl_dump_load {
 }
 pub(crate) use impl_dump_load;
 
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-pub struct IntegerBetween1And100(NonZeroU8);
-
-impl TryFrom<u64> for IntegerBetween1And100 {
-    type Error = &'static str;
-    fn try_from(data: u64) -> Result<Self, Self::Error> {
-        if data == 0 || data > 100 {
-            return Err("Invalid IntegerBetween1And100 value");
-        }
-
-        Ok(Self(NonZeroU8::new(data as u8).unwrap()))
-    }
-}
-
-impl From<IntegerBetween1And100> for u64 {
-    fn from(data: IntegerBetween1And100) -> Self {
-        u8::from(data.0) as u64
-    }
-}
-
 // This macro implements dump/load methods for client/server side.
 // It checks if both Req and Rep are implemented for a specified command
 // It also provides a way to use commands by specifying status, command and type.
@@ -61,11 +36,3 @@ impl From<IntegerBetween1And100> for u64 {
 parsec_protocol_cmds_family!("schema/invited_cmds");
 parsec_protocol_cmds_family!("schema/authenticated_cmds");
 parsec_protocol_cmds_family!("schema/anonymous_cmds");
-
-pub trait Request {
-    type Response: for<'de> Deserialize<'de>;
-
-    fn dump(self) -> Result<Vec<u8>, rmp_serde::encode::Error>;
-
-    fn load_response(buf: &[u8]) -> Result<Self::Response, ::rmp_serde::decode::Error>;
-}
