+++ oxidation/libparsec/crates/protocol/tests/test_ping.rs
@@ -2,8 +2,9 @@
 
 use hex_literal::hex;
 
+use libparsec_miniprotocol::Request;
 use libparsec_protocol::{
-    authenticated_cmds::v2 as authenticated_cmds, invited_cmds::v2 as invited_cmds, Request,
+    authenticated_cmds::v2 as authenticated_cmds, invited_cmds::v2 as invited_cmds,
 };
 use libparsec_tests_fixtures::parsec_test;
