+++ oxidation/libparsec/crates/serialization_format/Cargo.toml
@@ -15,7 +15,7 @@ quote = "1.0.26"
 syn = { version = "2.0.13", features = ["extra-traits"] }
 
 [dev-dependencies]
-libparsec_protocol = { path = "../protocol" }
+libparsec_miniprotocol = { path = "../miniprotocol" }
 libparsec_types = { path = "../types" }
 rstest = "0.17.0"
 pretty_assertions = "1.3.0"
